Fixing service shutdown logic to wait on requests instead of connections

+++ examples/chef/cookbooks/skynet-example/recipes/default.rb
@@ -27,7 +27,9 @@ execute "download-skynet" do
     go get github.com/bketelsen/skynet
   }
 
-  not_if "ls $GOPATH/src/github.com/bketelsen/skynet"
+  not_if do
+    File.exists?("/opt/local/gopath/src/github.com/bketelsen/skynet")
+  end
 end
 
 execute "update-skynet" do
@@ -48,6 +50,10 @@ execute "forced-rebuild" do
      rm fibservice service
    }
  end
+
+ not_if do
+   !File.exists?("/opt/local/gopath/bin") || !File.exists?("/opt/local/gopath/bin/service") || !File.exists?("/opt/local/gopath/bin/fibservice")
+ end
 end
 
 execute "install-example-service" do
@@ -57,7 +63,9 @@ execute "install-example-service" do
     go install  
   }
 
-  not_if "ls $GOPATH/bin/service"
+  not_if do
+    File.exists?("/opt/local/gopath/bin/service")
+  end
 end
 
 execute "install-fibonacci-service" do
@@ -67,5 +75,7 @@ execute "install-fibonacci-service" do
     go install  
   }
 
-  not_if "ls $GOPATH/bin/fibservice"
+  not_if do
+   File.exist?("/opt/local/gopath/bin/fibservice")
+  end
 end
